Btrfs: make avail_*_alloc_bits fields dynamic
Currently when new chunks are created respective avail_alloc_bits field
is updated to reflect profiles of all chunks present in the system.
However when chunks are removed, corresponding profile bits are never
cleared.

This patch clears corresponding bit of avail_alloc_bits field when the
last chunk of that type goes away.  Restriper needs this to properly
operate when "downgrading".

Signed-off-by: Ilya Dryomov <idryomov@gmail.com>