Btrfs: save restripe parameters to disk
Introduce a new btree objectid for storing restripe item.  The reason is
to be able to resume restriper after a crash with the same parameters.
Restripe item has a very high objectid and goes into tree of tree roots.

The key for the new item is as follows:

    [ BTRFS_RESTRIPE_OBJECTID ; BTRFS_RESTRIPE_ITEM_KEY ; 0 ]

Older kernels simply ignore it so it's safe to mount with an older
kernel and then go back to the newer one.

Signed-off-by: Ilya Dryomov <idryomov@gmail.com>