Btrfs: allow for pausing restriper
Implement an ioctl for pausing restriper.  This pauses the relocation,
but restripe is still considered to be "in progress": restriper item is
not deleted, other volume operations cannot be started, etc.  If paused
in the middle of profile changing operation we will continue making
allocations with the target profile.

Add a hook to close_ctree() to be able to pause restriper and free it's
data structures on unmount.  (It's safe to unmount when restriper is in
'paused' state, we will resume with the same parameters on the next
mount)

Signed-off-by: Ilya Dryomov <idryomov@gmail.com>