Btrfs: overwrite existing csums when logging
Users were complaining about the warning in btrfs_log_parent_inode() because
we're returning -EEXIST.  This warning is gone in another patch but it's
still making us commit a transaction when we don't have to.  Basically
because we will log all csums in an ordered extent for an em we have the
chance of logging the same csums multiple times.  Since it's not that big of
a deal just allow csums to be over written in logs and that way we are safe.
This gets rid of the EEXIST and we log like we normally would.  Thanks,

Signed-off-by: Josef Bacik <jbacik@fusionio.com>